DEXW-1982-asset-rates: move rates under matcher endpoint, add date q
