+++ src/endpoints/matcher/index.js
@@ -3,5 +3,8 @@ const Router = require('koa-router');
 const subrouter = new Router({ prefix: '/matcher/:matcher' });
 
 const pairs = require('./pairs');
+const rates = require('./rates');
 
-module.exports = pairsService => subrouter.use(pairs(pairsService).routes());
+module.exports = (pairsService, ratesService) => subrouter
+  .use(pairs(pairsService).routes())
+  .use(rates(ratesService).routes());
